feat: change queue implementation
