+++ engine/cdn/storage/storageunit_test.go
@@ -75,7 +75,7 @@ func TestRun(t *testing.T) {
 				},
 			},
 		},
-	}, storage.LogConfig{StepMaxSize: 3000, NbJobLogsGoroutines: 0, StepLinesRateLimit: 10})
+	})
 	require.NoError(t, err)
 	require.NotNil(t, cdnUnits)
 	cdnUnits.Start(ctx, sdk.NewGoRoutines())
@@ -113,11 +113,9 @@ func TestRun(t *testing.T) {
 	itemUnit, err = storage.LoadItemUnitByID(ctx, m, db, itemUnit.ID, gorpmapper.GetOptions.WithDecryption)
 	require.NoError(t, err)
 
-	_, err = cdnUnits.Buffer.Add(*itemUnit, 1.0, "this is the first log\n", storage.WithOption{IslastLine: false})
-	require.NoError(t, err)
+	require.NoError(t, cdnUnits.Buffer.Add(*itemUnit, 1.0, "this is the first log\n"))
 
-	_, err = cdnUnits.Buffer.Add(*itemUnit, 2.0, "this is the second log\n", storage.WithOption{IslastLine: false})
-	require.NoError(t, err)
+	require.NoError(t, cdnUnits.Buffer.Add(*itemUnit, 2.0, "this is the second log\n"))
 
 	reader, err := cdnUnits.Buffer.NewReader(context.TODO(), *itemUnit)
 	require.NoError(t, err)
